Consolidate TimeZoneInfo tests to a single file
